requirements.txt
absl-py==1.4.0
accelerate==0.22.0
aicrowd-api==0.1.23
aicrowd-gym==0.0.3
aiofiles==23.2.1
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.6
cloudpickle==2.2.1
cmake==3.26.4
colorlog==6.7.0
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.35
docker-pycreds==0.4.0
exceptiongroup==1.1.3
fastapi==0.101.1
faster-fifo==1.4.5
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gradio==3.41.1
gradio_client==0.5.0
grpcio==1.56.0
gym==0.23.0
gym-notices==0.0.8
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.7.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
line-profiler==4.0.3
lit==16.0.6
llvmlite==0.40.1
loguru==0.7.0
Markdown==3.4.3
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
minio==7.1.15
mpmath==1.3.0
msgpack==1.0.5
msgpack-numpy==0.4.8
networkx==3.1
nh3==0.2.14
ninja==1.11.1
nle==0.9.0
numba==0.57.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.131
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.1.2
oauthlib==3.2.2
opencv-python==4.7.0.72
orjson==3.9.5
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
peft==0.5.0
Pillow==8.2.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pydantic==1.10.12
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
ray==2.6.3
redis==4.5.5
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.28.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
termcolor==2.3.0
threadpoolctl==3.1.0
tiktoken==0.4.0
timeout-decorator==0.5.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
wandb==0.15.5
wavedrom==2.0.3.post3
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.6
zipp==3.15.0
zmq==0.0.0
-e ./rl_baseline/sample-factory
vllm==0.2.0
fschat==0.2.31
